+++ src/app/modules/app-routing.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core'
 import {RouterModule, Routes} from '@angular/router'
-import {DashboardLayoutComponent} from './dashboard-layout/dashboard-layout.component'
-import {LoginComponent} from './login/login.component'
+import {DashboardLayoutComponent} from '../components/dashboard-layout/dashboard-layout.component'
+import {LoginComponent} from '../pages/login/login.component'
 
 const routes: Routes = [
   {
